Updates to work with https://github.com/rails/rails/pull/44576
We can't call @raw_connection right away anymore, because it is nil until
it is used after the changes in https://github.com/rails/rails/pull/44576.

Also we need to have @lock established in order to be able to use
`with_raw_connection`. In the abstract_adapter implementation, that happens
after the arel_visitor is set, but in oracle enhanced, we need to use
`with_raw_connection` to check the database version in order to pick with
visitor to use. Calling lock_thread = nil before super means @lock isi
available to set the arel_visitor.

PR#44576 also requires an internal_exec_query method that is called by the
abstract adpater's exec_query. Naively renaming the oracle-enhanced's
exec_query to internal_exec_query makes the test pass, but there is probably
more to be done here.