Update HomogeneousIn for https://github.com/rails/rails/pull/45642
PR 45642 removes table_name and column_name which the existing
implementation relied upon. Change visit_Arel_Nodes_HomogeneousIn
to be closer the upstream implementation to implement the clause
with out trying to recreate the removed methods. Because of the
Oracle limit of 1000 values in an IN clause we can't use it directly
but we can use a very similar approach to the upstream version,
calling a method very close to the upstream version for each 1000
values and joining them together with an OR or AND.

The upstream_visit_Arel_Nodes_HomogeneousIn_for_values is really
close to the visit_Arel_Nodes_HomogeneousIn at the time of PR 45642.
It would be nice to make it match exactly but we need to give the
bind parameter a name starting with `:` rather than a bind called
`?` which is what the default code gives us. The method signature
also needs to change so it can receive the values as we slice the
IN clause that has more than 1000 values.

NOTE: this changes the name of the bind by passing the optional bind_block,
but I wonder if the same thing could be achieved using o.proc_for_binds.