More support for rails #44576
https://github.com/rails/rails/pull/44576 adds `with_raw_connection`. Using
that in database_statements instead of @raw_connection mostly works, except
 it broke 2 tests:
* ["should reconnect and execute query if connection is lost and auto retry is enabled"](https://github.com/rsim/oracle-enhanced/blob/deb214decc3799608c8be386e91c6c7531c59793/spec/active_record/connection_adapters/oracle_enhanced/connection_spec.rb#L493)
* ["among other things will reconnect and retry if there is a connection error.](https://github.com/rsim/oracle-enhanced/blob/deb214decc3799608c8be386e91c6c7531c59793/spec/active_record/connection_adapters/oracle_enhanced/connection_spec.rb#L500)

`with_raw_connection` has a retry and reconnect built into it, but it was
catching the oracle connection error. Adjust `retryable_connection_error?`
so that both is aware of the OCI errors and reflects the `should_retry?`
logic of the `with_retry` that database_statements used before.